handle the biginteger conversion

+++ airbyte-cdk/bulk/toolkits/load-iceberg-parquet/src/main/kotlin/io/airbyte/cdk/load/data/iceberg/parquet/AirbyteValueToIcebergRecord.kt
@@ -53,7 +53,7 @@ class AirbyteValueToIcebergRecord {
             is BooleanValue -> return airbyteValue.value
             is DateValue ->
                 throw IllegalArgumentException("String-based date types are not supported")
-            is IntegerValue -> return airbyteValue.value
+            is IntegerValue -> return airbyteValue.value.toLong()
             is IntValue -> return airbyteValue.value
             is NullValue -> return null
             is NumberValue -> return airbyteValue.value.toDouble()
